requirements.txt
openai==1.25.1
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.31.1
markdown==3.6
